example big numbers for exam

+++ C02/ex03/src/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: cariencaljouw <cariencaljouw@student.co      +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/08/23 14:31:27 by cariencaljo   #+#    #+#                 */
-/*   Updated: 2023/09/06 11:18:39 by ccaljouw      ########   odam.nl         */
+/*   Updated: 2023/09/15 08:35:56 by cariencaljo   ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -53,7 +53,7 @@ void	Fixed::setRawBits(int const raw) {
 }
 
 float	Fixed::toFloat(void) const {
-	float f = ((float)(this->getRawBits()) / (float)(1 << _fractionalBits));
+	float f = (float)(this->getRawBits()) / (1 << _fractionalBits);
 	return f;
 }
